move parse_js_request_body to L8BodyType.from_jsvalue
