+++ src/types/request/mod.rs
@@ -1,13 +1,18 @@
+mod body;
+mod mode_and_policies;
+
 use wasm_bindgen::{JsCast, JsValue, throw_str, UnwrapThrowExt};
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use ntor::common::NTorParty;
-use web_sys::{ReferrerPolicy, RequestMode};
 use web_sys::{AbortSignal, console, Request, RequestInit, ResponseInit};
+use body::L8BodyType;
+use mode_and_policies::L8RequestMode;
 use crate::storage::InMemoryCache;
 use crate::types::response::L8ResponseObject;
 use crate::types::{network_state::NetworkStateOpen, WasmEncryptedMessage};
 use crate::types::network_state::NetworkStateResponse;
+use crate::types::request::mode_and_policies::get_request_referer_policy;
 use crate::utils;
 
 /// A JSON serializable wrapper for a request that can be sent using the Fetch API.
@@ -91,7 +96,7 @@ impl L8RequestObject {
 
         let body = options.get_body();
         if !body.is_undefined() && !body.is_null() {
-            let body = utils::parse_js_request_body(body).await.map_err(|e| {
+            let body = L8BodyType::from_jsvalue(body).await.map_err(|e| {
                 JsValue::from_str(&format!(
                     "Failed to parse request body: {}",
                     e.as_string().unwrap_or_else(|| "Unknown error".to_string())
@@ -123,8 +128,7 @@ impl L8RequestObject {
                         serde_json::to_value(&format!(
                             "multipart/form-data; boundary={}",
                             boundary
-                        ))
-                            .expect_throw("a valid string is JSON serializable"),
+                        )).expect_throw("a valid string is JSON serializable"),
                     );
 
                     req_wrapper.body = data;
@@ -386,13 +390,7 @@ impl L8RequestObject {
             .inspect(|v| self.keep_alive = Some(*v));
 
         // mode
-        self.mode = match options.get_mode() {
-            Some(RequestMode::SameOrigin) => Some(L8RequestMode::SameOrigin),
-            Some(RequestMode::NoCors) => Some(L8RequestMode::NoCors),
-            Some(RequestMode::Cors) => Some(L8RequestMode::Cors),
-            Some(RequestMode::Navigate) => Some(L8RequestMode::Navigate),
-            _ => Some(L8RequestMode::Cors),
-        };
+        self.mode = L8RequestMode::from_request_options(options);
 
         // redirect
         _ = js_sys::Reflect::get(&options, &"redirect".into()).inspect(|v| {
@@ -401,20 +399,7 @@ impl L8RequestObject {
         });
 
         // referrer policy
-        let mut referrer_policy = "";
-        if let Some(referrer_policy_) = options.get_referrer_policy() {
-            referrer_policy = match referrer_policy_ {
-                ReferrerPolicy::NoReferrer => "no-referrer",
-                ReferrerPolicy::NoReferrerWhenDowngrade => "no-referrer-when-downgrade",
-                ReferrerPolicy::Origin => "origin",
-                ReferrerPolicy::OriginWhenCrossOrigin => "origin-when-cross-origin",
-                ReferrerPolicy::UnsafeUrl => "unsafe-url",
-                ReferrerPolicy::SameOrigin => "same-origin",
-                ReferrerPolicy::StrictOrigin => "strict-origin",
-                ReferrerPolicy::StrictOriginWhenCrossOrigin => "strict-origin-when-cross-origin",
-                _ => "",
-            };
-        }
+        let referrer_policy = get_request_referer_policy(options);
 
         if !referrer_policy.is_empty() {
             self.headers.insert(
@@ -441,27 +426,4 @@ impl L8RequestObject {
         // signal
         self.signal = options.get_signal();
     }
-}
-
-pub enum L8BodyType {
-    Bytes(Vec<u8>),
-    Stream(wasm_streams::ReadableStream),
-    Params(HashMap<String, String>),
-    FormData(web_sys::FormData),
-    #[allow(dead_code)]
-    File(web_sys::File),
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub enum L8RequestMode {
-    // Disallows cross-origin requests. If a request is made to another origin with this mode set, the result is an error.
-    SameOrigin = 0,
-    // Disables CORS for cross-origin requests. The response is opaque, meaning that its headers and body are not available to JavaScript.
-    NoCors = 1,
-    // If the request is cross-origin then it will use the Cross-Origin Resource Sharing (CORS) mechanism.
-    // Using the Request() constructor, the value of the mode property for that Request is set to cors.
-    Cors = 2,
-    // A mode for supporting navigation. The navigate value is intended to be used only by HTML navigation.
-    // A navigate request is created only while navigating between documents.
-    Navigate = 3,
 }
